Move gridAtomics to Grid file
