+++ packages/pigment-css-react/src/Grid.jsx
@@ -4,7 +4,7 @@ import clsx from 'clsx';
 import PropTypes from 'prop-types';
 import * as React from 'react';
 
-import { gridAtomics } from './baseAtomics';
+import { generateAtomics } from './generateAtomics';
 import styled from './styled';
 
 function isGridComponent(element) {
@@ -14,7 +14,85 @@ function isGridComponent(element) {
   return element.type.muiName === 'Grid' || element.type?._payload?.value?.muiName === 'Grid';
 }
 
+export const gridAtomics = generateAtomics(({ theme }) => ({
+  conditions: Object.keys(theme.breakpoints.values).reduce((acc, breakpoint) => {
+    acc[breakpoint] = `@media (min-width: ${theme.breakpoints.values[breakpoint]}${
+      theme.breakpoints.unit ?? 'px'
+    })`;
+    return acc;
+  }, {}),
+  defaultCondition: theme.breakpoints?.keys?.[0] ?? 'xs',
+  properties: {
+    flexDirection: ['column', 'column-reverse', 'row', 'row-reverse'],
+    '--Grid-parent-column-count': ['--Grid-parent-column-count'],
+    '--Grid-parent-column-spacing': ['--Grid-parent-column-spacing'],
+    '--Grid-parent-row-spacing': ['--Grid-parent-row-spacing'],
+    '--Grid-self-column-span': ['--Grid-self-column-span'],
+    '--Grid-self-width': ['--Grid-self-width'],
+    '--Grid-self-max-width': ['--Grid-self-max-width'],
+    '--Grid-self-flex': ['--Grid-self-flex'],
+    '--Grid-self-column-spacing': ['--Grid-self-column-spacing'],
+    '--Grid-self-row-spacing': ['--Grid-self-row-spacing'],
+    '--Grid-self-offset': ['--Grid-self-offset'],
+    '--Grid-self-margin-left': ['--Grid-self-margin-left'],
+  },
+  shorthands: {
+    direction: ['flexDirection'],
+  },
+  unitless: ['--Grid-parent-column-count', '--Grid-self-column-span', '--Grid-self-offset'],
+  multipliers: {
+    '--Grid-parent-column-spacing': Array.isArray(theme.vars?.spacing) ? theme.vars.spacing[0] : theme.vars?.spacing,
+    '--Grid-parent-row-spacing': Array.isArray(theme.vars?.spacing) ? theme.vars.spacing[0] : theme.vars?.spacing,
+    '--Grid-self-column-spacing': Array.isArray(theme.vars?.spacing) ? theme.vars.spacing[0] : theme.vars?.spacing,
+    '--Grid-self-row-spacing': Array.isArray(theme.vars?.spacing) ? theme.vars.spacing[0] : theme.vars?.spacing,
+  },
+  inlineGetters: {
+    '--Grid-self-width': (value) => {
+      if (value === 'grow') {
+        return 'unset';
+      }
+
+      if (value === 'auto') {
+        return 'auto';
+      }
+
+      return 'var(--Grid-fixed-width)';
+    },
+    '--Grid-self-max-width': (value) => {
+      if (value === 'grow') {
+        return '100%';
+      }
+
+      if (value === 'auto') {
+        return 'none';
+      }
+
+      return 'unset';
+    },
+    '--Grid-self-flex': (value) => {
+      if (value === 'grow') {
+        return '1 1 0';
+      }
+
+      if (value === 'auto') {
+        return '0 0 auto';
+      }
+
+      return '0 1 auto';
+    },
+    '--Grid-self-margin-left': (value) => {
+      if (value === 'auto') {
+        return 'auto';
+      }
+
+      return 'var(--Grid-fixed-offset)';
+    },
+  },
+}));
+
 const GridComponent = styled('div')({
+  '--Grid-fixed-width': 'calc(100% * var(--Grid-self-column-span) / var(--Grid-parent-column-count) - (var(--Grid-parent-column-count) - var(--Grid-self-column-span)) * var(--Grid-parent-column-spacing) / var(--Grid-parent-column-count))',
+  '--Grid-fixed-offset': 'calc(100% * var(--Grid-self-offset) / var(--Grid-parent-column-count) + var(--Grid-parent-column-spacing) * var(--Grid-self-offset) / var(--Grid-parent-column-count))',
   variants: [
     {
       props: { container: true },
